requirements.txt
google-generativeai==0.8.3
groq==0.11.0
gspread==6.1.2
google-auth==2.35.0
python-dotenv==1.0.1
sib-api-v3-sdk==7.6.0
schedule==1.2.0
razorpay==2.0.0
twilio==9.10.2
fastapi==0.115.8
sqlalchemy==1.4.54
uvicorn==0.34.0
requests==2.32.3
jinja2==3.1.5
python-multipart==0.0.20